a new hook to split conf on multifilters
